worker/pyproject.toml
[project]
name = "stroke-pred-api"
version = "0.1.0"
description = "Stroke Prediction API for Cloudflare Workers"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.115.12",
    "pydantic==2.10.6",
    "workers-asgi @ git+https://github.com/cloudflare/workers-asgi.git@main"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.wrangler]
name = "stroke-pred-api"
main = "ml_api_stroke.py"
compatibility_date = "2025-03-14"